+++ dioptas/model/util/HelperModule.py
@@ -1,0 +1,214 @@
+# -*- coding: utf8 -*-
+# Dioptas - GUI program for fast processing of 2D X-ray data
+# Copyright (C) 2014  Clemens Prescher (clemens.prescher@gmail.com)
+# GSECARS, University of Chicago
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+__author__ = 'Clemens Prescher'
+
+import os
+import re
+import time
+
+import numpy as np
+from PyQt4 import QtCore
+from colorsys import hsv_to_rgb
+
+
+class FileNameIterator(QtCore.QObject):
+    # TODO create an File Index and then just get the next files according to this.
+    # Otherwise searching a network is always to slow...
+
+    def __init__(self, filename=None):
+        super(FileNameIterator, self).__init__()
+        self.acceptable_file_endings = []
+        self.directory_watcher = QtCore.QFileSystemWatcher()
+        self.directory_watcher.directoryChanged.connect(self.add_new_files_to_list)
+        self.create_timed_file_list = False
+
+        if filename is None:
+            self.complete_path = None
+            self.directory = None
+            self.filename = None
+            self.file_list = []
+            self.ordered_file_list = []
+            self.filename_list = []
+        else:
+            self.complete_path = os.path.abspath(filename)
+            self.directory, self.filename = os.path.split(self.complete_path)
+            self.acceptable_file_endings.append(self.filename.split('.')[-1])
+
+    def _get_files_list(self):
+        t1 = time.time()
+        filename_list = os.listdir(self.directory)
+        files = []
+        for file in filename_list:
+            if self.is_correct_file_type(file):
+                files.append(file)
+        paths = [os.path.join(self.directory, file) for file in files]
+        file_list = [(os.path.getctime(path), path) for path in paths]
+        self.filename_list = paths
+        print('Time needed  for getting files: {0}s.'.format(time.time() - t1))
+        return file_list
+
+    def is_correct_file_type(self, filename):
+        for ending in self.acceptable_file_endings:
+            if filename.endswith(ending):
+                return True
+        return False
+
+    def _order_file_list(self):
+        t1 = time.time()
+        self.ordered_file_list = self.file_list
+        self.ordered_file_list.sort(key=lambda x: x[0])
+
+        print('Time needed  for ordering files: {0}s.'.format(time.time() - t1))
+
+    def update_file_list(self):
+        self.file_list = self._get_files_list()
+        self._order_file_list()
+
+    def _iterate_file_number(self, path, step):
+        directory, file_str = os.path.split(path)
+        pattern = re.compile(r'\d+')
+
+        match_iterator = pattern.finditer(file_str)
+
+        for match in reversed(list(match_iterator)):
+            number_span = match.span()
+            left_ind = number_span[0]
+            right_ind = number_span[1]
+            number = int(file_str[left_ind:right_ind]) + step
+            new_file_str = "{left_str}{number:0{len}}{right_str}".format(
+                left_str=file_str[:left_ind],
+                number=number,
+                len=right_ind - left_ind,
+                right_str=file_str[right_ind:]
+            )
+            new_complete_path = os.path.join(directory, new_file_str)
+            if os.path.exists(new_complete_path):
+                self.complete_path = new_complete_path
+                return new_complete_path
+        return None
+
+    def get_next_filename(self, step=1, filename=None, mode='number'):
+        if filename is not None:
+            self.complete_path = filename
+
+        if self.complete_path is None:
+            return None
+        if mode == 'time':
+            time_stat = os.path.getctime(self.complete_path)
+            cur_ind = self.ordered_file_list.index((time_stat, self.complete_path))
+            # cur_ind = self.ordered_file_list.index(self.complete_path)
+            try:
+                self.complete_path = self.ordered_file_list[cur_ind + step][1]
+                return self.complete_path
+            except IndexError:
+                return None
+        elif mode == 'number':
+            return self._iterate_file_number(self.complete_path, step)
+
+    def get_previous_filename(self, step=1, mode='number'):
+        """
+        Tries to get the previous filename.
+
+        :param mode:
+            can have two values either number or mode. Number will decrement the last digits of the file name \
+            and time will get the next file by creation time.
+        :return:
+            either new filename as a string if it exists or None
+        """
+        if self.complete_path is None:
+            return None
+        if mode == 'time':
+            time_stat = os.path.getctime(self.complete_path)
+            cur_ind = self.ordered_file_list.index((time_stat, self.complete_path))
+            # cur_ind = self.ordered_file_list.index(self.complete_path)
+            if cur_ind > 0:
+                try:
+                    self.complete_path = self.ordered_file_list[cur_ind - step][1]
+                    return self.complete_path
+                except IndexError:
+                    return None
+        elif mode == 'number':
+            return self._iterate_file_number(self.complete_path, -step)
+
+    def update_filename(self, new_filename):
+        self.complete_path = os.path.abspath(new_filename)
+        new_directory, file_str = os.path.split(self.complete_path)
+        try:
+            self.acceptable_file_endings.append(file_str.split('.')[-1])
+        except AttributeError:
+            pass
+        if self.directory != new_directory:
+            if self.directory is not None:
+                self.directory_watcher.removePath(self.directory)
+            self.directory_watcher.addPath(new_directory)
+            self.directory = new_directory
+            if self.create_timed_file_list:
+                self.update_file_list()
+
+        if (self.create_timed_file_list and self.ordered_file_list == []):
+            self.update_file_list()
+
+    def add_new_files_to_list(self):
+        """
+        checks for new files in folder and adds them to the sorted_file_list
+        :return:
+        """
+        cur_filename_list = os.listdir(self.directory)
+        cur_filename_list = [os.path.join(self.directory, filename) for filename in cur_filename_list if
+                             self.is_correct_file_type(filename)]
+        new_filename_list = [filename for filename in cur_filename_list if filename not in list(self.filename_list)]
+        self.filename_list = cur_filename_list
+        for filename in new_filename_list:
+            creation_time = os.path.getctime(filename)
+            if len(self.ordered_file_list) > 0:
+                if creation_time > self.ordered_file_list[-1][0]:
+                    self.ordered_file_list.append((creation_time, filename))
+                else:
+                    for ind in xrange(len(self.ordered_file_list)):
+                        if creation_time < self.ordered_file_list[ind][0]:
+                            self.ordered_file_list.insert(ind, (creation_time, filename))
+                            break
+            else:
+                self.ordered_file_list.append((creation_time, filename))
+
+
+def rotate_matrix_m90(matrix):
+    return np.rot90(matrix, -1)
+
+
+def rotate_matrix_p90(matrix):
+    return np.rot90(matrix)
+
+
+def get_base_name(filename):
+    str = os.path.basename(filename)
+    if '.' in str:
+        str = str.split('.')[:-1][0]
+    return str
+
+
+def calculate_color(ind):
+    s = 0.8
+    v = 0.8
+    h = (0.19 * (ind + 2)) % 1
+    return np.array(hsv_to_rgb(h, s, v)) * 255
+
+
+def convert_d_to_two_theta(d, wavelength):
+    return np.arcsin(wavelength / (2 * d)) / np.pi * 360